python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ecf-dgii"
version = "0.3.0"
description = "Python SDK for the ECF DGII API (Dominican Republic Electronic Fiscal Receipts)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "SSD Smart Software Development", email = "contacto@ssd.com.do" },
]
keywords = ["dgii", "ecf", "fiscal", "dominican-republic", "electronic-invoice", "tax"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27,<1",
    "attrs>=23.1.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/SSD-Smart-Software-Development-SRL/ecf_dgii"
Documentation = "https://github.com/SSD-Smart-Software-Development-SRL/ecf_dgii/tree/main/python#readme"
Repository = "https://github.com/SSD-Smart-Software-Development-SRL/ecf_dgii"
Issues = "https://github.com/SSD-Smart-Software-Development-SRL/ecf_dgii/issues"

[tool.hatch.build.targets.wheel]
packages = ["ecf_dgii"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
